Logs Icon Properly Opens Logs
Updated the logs icon for both the desktop and taskbar to open the logs themself after the correct password has been entered. Created a Computer Icon script that is currently only used by the Logs Icon and update the Password Enter to include information about what type of screen is being opened, as to better communicate with  other scripts.